[Code] Fixed some compiling issues.
Issue #20.
Now reads:
[14:40:45] Compiling act_move.c ...
act_move.c: In function ‘find_path’:
act_move.c:2005: warning: cast to pointer from integer of different size

+++ src/act_move.c
@@ -35,6 +35,8 @@
 #include <malloc.h>
 #include <stdlib.h>
 #include <time.h>
+#include <stdint.h>
+
 #include "twilight.h"
 #include "tables.h"
 #include "lookup.h"
@@ -1998,7 +2000,7 @@ int find_path( int in_room_vnum, int out_room_vnum, CHAR_DATA *ch, int depth, in
 
                             /* ancestor for first layer is the direction */
                             hash_enter( &x_room, tmp_room,
-                                    ((int)hash_find(&x_room,q_head->room_nr)
+                                    ((uintptr_t)hash_find(&x_room,q_head->room_nr)
                                             == -1) ? (void*)(i+1)
                                                     : hash_find(&x_room,q_head->room_nr));
                         }
@@ -2013,7 +2015,7 @@ int find_path( int in_room_vnum, int out_room_vnum, CHAR_DATA *ch, int depth, in
                             free(q_head);
                         }
                         /* return direction if first layer */
-                        if ((int)hash_find(&x_room,tmp_room)==-1)
+                        if ((uintptr_t)hash_find(&x_room,tmp_room)==-1)
                         {
                             if (x_room.buckets)
                             {
@@ -2027,7 +2029,7 @@ int find_path( int in_room_vnum, int out_room_vnum, CHAR_DATA *ch, int depth, in
                             /* else return the ancestor */
                             int i;
 
-                            i = (int)hash_find(&x_room,tmp_room);
+                            i = (uintptr_t)hash_find(&x_room,tmp_room);
                             if (x_room.buckets)
                             {
                                 /* junk left over from a previous track */
